cycloid-toolkit change vault-approle terraform format
Change the default yaml format used by terraform var-file
to the terraform tfvars format.
This allow terraform to automatically load this file if it is present in the terraform directory.

(require to be in terraform vars format `foo = "bar"`. And to stay compatible with the var-file in the commandline,
the file have to be named with the tfvars extention. If not terraform will try to load it as yaml and fail.